+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email 'isuftin@usgs.gov'
 license          'CPL-1.0'
 description      'Installs/Configures CIS STIG benchmarks'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.6.9'
+version          '0.6.10'
 source_url		 'https://github.com/USGS-CIDA/stig'
 issues_url		 'https://github.com/USGS-CIDA/stig/issues'